Time: 856 ms (66.67%) | Memory: 94 MB (40.00%) - LeetSync

+++ 3699-count-special-subsequences/count-special-subsequences.cpp
@@ -6,16 +6,16 @@ public:
         long long ans = 0;
         for(int i = 0; i < n - 6; i++){
             for(int j = i+2; j < n - 4; j++){
-                mp[(double) nums[i] / (double) nums[j]]++;
+                mp[(double) nums[i] /nums[j]]++;
             }
         }
 
         for(int i = n - 3; i >= 4; i--){
             for(int j = i+2; j < n; j++){
-                ans += mp[(double) nums[j] / (double) nums[i]];
+                ans += mp[(double) nums[j] /nums[i]];
             }
             for(int j = 0; j <= i - 4; j++){
-                mp[(double) nums[j] / (double) nums[i-2]]--;
+                mp[(double) nums[j] /nums[i-2]]--;
             }
         }
         return ans;
